refactor: extract events formatting to their own file
